requirements.txt
pandas
pyarrow
PyYAML
scikit-learn
joblib
